+++ NEWS.md
@@ -1,5 +1,8 @@
 # herehelper 0.1.0.900
 
+* Support for cleaning `file.path()` calls was added to `here_clean_path()`. The function calls to be targeted for cleaning can be controlled using the `fns` argument.
+* A new function, `here_clean_all()` was added to clean all `here()` and `file.path()` calls in the active document, rather than those highlighted. The function provides a print out of the number of each call it has found and cleaned.
+
 * Initial working version of the package.
 * Added a `NEWS.md` file to track changes to the package.